+++ ElectronicObserver.KancolleApi.Types/Models/APIAirBaseAttackAPIStage3.cs
@@ -15,7 +15,7 @@ public class ApiAirBaseAttackApiStage3
 	[JsonPropertyName("api_edam")]
 	[JsonIgnore(Condition = JsonIgnoreCondition.Never)]
 	[Required]
-	public List<int> ApiEdam { get; set; } = new();
+	public List<double> ApiEdam { get; set; } = new();
 
 	[JsonPropertyName("api_erai_flag")]
 	[JsonIgnore(Condition = JsonIgnoreCondition.Never)]
